Convert amount to BigDecimal type for Invoice

+++ src/main/java/com/sms/invoicify/models/InvoiceDto.java
@@ -7,6 +7,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -22,5 +23,5 @@ public class InvoiceDto {
   private List<Item> items;
   private String companyName;
   private PaymentStatus paymentStatus;
-  private Double totalCost;
+  private BigDecimal totalCost;
 }
